impl/windows/windows.js to impl/windows/index.js
